+++ Changelog.md
@@ -57,3 +57,4 @@ site itself.
 # Version 0.4.1
 - Bug fix for `cmv` so that `jamaly` r_corr is always positive. 
 - Demo for digitizing a plant equipment layout added by @williamhobbs
+- Improve comments in `cmv_demo`